# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sun Glass Accesorios - E-Commerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: products
    description: Product management
  - name: categories
    description: Category listing

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, name_az]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of category names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /products/stats:
    get:
      operationId: getProductStats
      tags: [products]
      summary: Get product statistics (admin dashboard)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Product stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStats"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthToken:
      type: object
      required: [token]
      properties:
        token:
          type: string

    Product:
      type: object
      required: [id, name, category, price, description, image_url, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        description:
          type: string
        image_url:
          type: string
        created_at:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"

    ProductInput:
      type: object
      required: [name, category, price, description, image_url]
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        description:
          type: string
        image_url:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantInput"

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        description:
          type: string
        image_url:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantInput"

    ProductVariant:
      type: object
      required: [id, name, available]
      properties:
        id:
          type: integer
        name:
          type: string
        available:
          type: boolean

    ProductVariantInput:
      type: object
      required: [name, available]
      properties:
        id:
          type: integer
        name:
          type: string
        available:
          type: boolean

    ProductStats:
      type: object
      required: [totalProducts, totalCategories, byCategory, priceRange]
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        priceRange:
          $ref: "#/components/schemas/PriceRange"

    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer

    PriceRange:
      type: object
      required: [min, max, avg]
      properties:
        min:
          type: number
        max:
          type: number
        avg:
          type: number

    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string