# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sun Glass Accesorios - E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariant } from './productVariant';

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  description: string;
  image_url: string;
  created_at: string;
  variants?: ProductVariant[];
}