# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sun Glass Accesorios - E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariantInput } from './productVariantInput';

export interface ProductInput {
  name: string;
  category: string;
  price: number;
  description: string;
  image_url: string;
  variants?: ProductVariantInput[];
}