# lib/api-zod/src/generated/types/productVariant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sun Glass Accesorios - E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductVariant {
  id: number;
  name: string;
  available: boolean;
}